factory file hierarchy refactored
