+++ src/factory/util.ts
@@ -1,0 +1,40 @@
+import moment from 'moment';
+import request from 'request';
+
+import { Answer } from '../models/Questionary';
+import { DataType } from '../models/Template';
+import { logger } from '../utils/Logger';
+
+export const getFileAttachmentIds = (answer: Answer) => {
+  if (answer.question.dataType === DataType.FILE_UPLOAD && answer.value) {
+    if (Array.isArray(answer.value)) {
+      return answer.value;
+    } else if (typeof answer.value === 'string') {
+      return answer.value.split(',');
+    } else {
+      logger.logError(
+        'Questionary answer with DataType `FILE_UPLOAD` has neither string nor array value',
+        { answer }
+      );
+
+      return [];
+    }
+  }
+
+  return [];
+};
+
+export const bufferRequestBody = (req: request.Request) =>
+  new Promise(resolve => {
+    const buffer: Buffer[] = [];
+
+    req.on('data', chunk =>
+      buffer.push(typeof chunk === 'string' ? Buffer.from(chunk) : chunk)
+    );
+
+    req.on('complete', () => resolve(Buffer.concat(buffer).toString()));
+  });
+
+export function getCurrentTimestamp() {
+  return moment.utc().format('YYYY-MM-DD_hhmmss');
+}
